Implemented delta time

+++ 2D_GameEngine/src/Game.h
@@ -8,7 +8,7 @@ class Game
 {
 private:
 	bool isRunning;
-	int millisecsPreviousFrame = 0;
+	double millisecsPreviousFrame = 0;
 	SDL_Window* window;
 	SDL_Renderer* renderer;
